Fixed various ArrayIndexOutOfBoundsExceptions in the code.
Improved the behavior of WicketIdGotoAction so that it falls back appropriately if the user clicks on a non-Wicket element in the source file.  I think it's fixed.

Added the WicketIdCompletionComponent and WicketJavaIdCompletionData classes.  These handle autocompletion in Wicket markup files.  

The changes tested fine on IDEA 6, but it's my first bit of plugin code, so it's anybody's guess if it'll work elsewhere.

git-svn-id: file:///home/igor/dev/stuff/svnbackup/trunk@1473 ef7698a4-5110-0410-9fc6-c7eb3693863f